mm/hmm: change hmm_vma_fault() to allow write fault on page basis
Change hmm_vma_fault() to not take a global write fault flag for a range
but instead rely on the caller to populate HMM pfns array with proper
fault flag ie HMM_PFN_VALID if the driver wants a read fault for that
address or HMM_PFN_VALID and HMM_PFN_WRITE for write.

Moreover by setting HMM_PFN_DEVICE_PRIVATE the device driver can ask for
device private memory to be migrated back to system memory through page
faults.

This is a more flexible API and it better reflects how a device handles
and reports faults.

Link: http://lkml.kernel.org/r/20180316203552.4155-4-jglisse@redhat.com
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
Cc: Evgeny Baskakov <ebaskakov@nvidia.com>
Cc: Ralph Campbell <rcampbell@nvidia.com>
Cc: Mark Hairgrove <mhairgrove@nvidia.com>
Cc: John Hubbard <jhubbard@nvidia.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>