lib/reed_solomon: remove VLAs by setting upper bound on nroots
Avoid stack VLAs[1] by always allocating the upper bound of stack space
needed.  The existing users of rslib appear to max out at 24 roots[2], so
use that as the upper bound until we have a reason to change it.

Alternative considered: make init_rs() a true caller-instance and
pre-allocate the workspaces.  This would possibly need locking and a
refactoring of the returned structure.

Using kmalloc in this path doesn't look great, especially since at least
one caller (pstore) is sensitive to allocations during rslib usage (it
expects to run it during an Oops, for example).

[1] https://lkml.org/lkml/2018/3/7/621
[2] https://lkml.org/lkml/2018/3/9/838

Link: http://lkml.kernel.org/r/20180315225919.GA43806@beast
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Segher Boessenkool <segher@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>