zram: correct flag name of ZRAM_ACCESS
Patch series "zram memory tracking", v5.

zRam as swap is useful for small memory device.  However, swap means those
pages on zram are mostly cold pages due to VM's LRU algorithm. 
Especially, once init data for application are touched for launching, they
tend to be not accessed any more and finally swapped out.  zRAM can store
such cold pages as compressed form but it's pointless to keep in memory. 
As well, it's pointless to store incompressible pages to zram so better
idea is app developers manages them directly like free or mlock rather
than remaining them on heap.

This patch provides a debugfs /sys/kernel/debug/zram/zram0/block_state to
represent each block's state so admin can investigate what memory is
cold|incompressible|same page with using pagemap once the pages are
swapped out.

The output is as follows:
      300    75.033841 .wh
      301    63.806904 s..
      302    63.806919 ..h

First column is zram's block index and 3rh one represents symbol (s: same
page w: written page to backing store h: huge page) of the block state. 
Second column represents usec time unit of the block was last accessed. 
So above example means the 300th block is accessed at 75.033851 second and
it was huge so it was written to the backing store.


This patch (of 4):

ZRAM_ACCESS is used for locking a slot of zram so correct the name.  It is
also not a common flag to indicate status of the block so move the declare
position on top of the flag.  Lastly, let's move the function to the top
of source code to be able to use it easily without forward declaration.

Link: http://lkml.kernel.org/r/20180416090946.63057-2-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>