lib/rhashtable: reorder some inititalization sequences
rhashtable_init() allocates memory at the very end of the call, once
everything is setup; with the exception of the nelems parameter.  However,
unless the user is doing something bogus with params for which -EINVAL is
returned, memory allocation is the only operation that can trigger the
call to fail.

Thus move bucket_table_alloc() up such that we fail back to the caller
asap, instead of doing useless checks.  This is safe as the the table
allocation isn't using the halfly setup 'ht' structure and
bucket_table_alloc() call chain only ends up using the ht->nulls_base
member in INIT_RHT_NULLS_HEAD.

Also move the locking initialization down to the end.

Link: http://lkml.kernel.org/r/20180514151332.31352-1-dave@stgolabs.net
Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Thomas Graf <tgraf@suug.ch>
Cc: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>